Clarified presentation download directory name. Also if no name is provided then search keywords sepearted with a '_'will be used as a directory name.
Additionally, a new cache system is set up. First, we will get all the pptx urls and then before downloading, we filter those that are already cached, and load only those that are not cached.

Lastly, before relying on the cache, we program loads a "cleaned up" cache before any operation on the cache. The clean up will ensure that all the entries within the cache point to a valid file.